Refactor ReadingPageViewModel to have one base view model and a separate subclass that does the main reading page.
Also start building data structures for updated question searching API.
