+++ RightToAskClient/RightToAskClient/RightToAskClient/Models/ServerCommsData/Page.cs
@@ -1,0 +1,15 @@
+using System.Text.Json.Serialization;
+
+// For pagination when requesting questions from the server.
+namespace RightToAskClient.Models.ServerCommsData
+{
+    public class Page
+    {
+        [JsonPropertyName("from")] 
+        public int from { get; set; } 
+        
+        [JsonPropertyName("to")] 
+        public int to { get; set; } 
+        
+    }
+}
