+++ RightToAskClient/RightToAskClient/RightToAskClient/Resources/serverconfig.json
@@ -1,6 +1,5 @@
 {
-  "remoteServerUse":true,
+  "remoteServerUse":false,
   "url":"https://righttoasktestserver2.democracydevelopers.org.au",
   "remoteServerPublicKey":"z5iEoNI7T/vEnFMkVK/jCj/NCT/K+0/J3jigYdwR6Yw=",
-  "localServerPublicKey":""
-}
+  "localServerPublicKey":"OJ/tBn4rOrOebgbICBi3i2oflO0hqz0D8daItDZ53vI="}
